+++ dart/geobase/lib/coordinates.dart
@@ -18,6 +18,7 @@ export 'src/codes/axis_order.dart';
 export 'src/codes/cardinal_precision.dart';
 export 'src/codes/coords.dart';
 export 'src/codes/dms_type.dart';
+export 'src/codes/geo_representation.dart';
 export 'src/constants/epsilon.dart';
 export 'src/coordinates/base/aligned.dart';
 export 'src/coordinates/base/box.dart';