+++ dart/geobase/lib/geobase.dart
@@ -28,6 +28,7 @@ export 'src/codes/canvas_origin.dart';
 export 'src/codes/cardinal_precision.dart';
 export 'src/codes/coords.dart';
 export 'src/codes/dms_type.dart';
+export 'src/codes/geo_representation.dart';
 export 'src/codes/geom.dart';
 
 // constants